Move JWT service token strategy out of middleware
